backend/requirements.txt
flask
flask-cors
pillow
opencv-python
numpy
twilio
requests
python-dotenv
